🔥 chore(linting): remove unused eslintignore files and update eslint-config-polished
The files .eslintignore in the root directory and in the directories apps/space-explorer-api and apps/space-explorer-ui were removed as they were not being used. The file packages/eslint-config-polished/javascript.js was updated to include the ignorePatterns property, which specifies the folders to be ignored by eslint. The ignorePatterns property was added to improve the linting process by ignoring specific folders and allowing dotfiles to be linted.